multinest: Add scripts for prior transformation.
Add a mathematica script to which prints the integration values needed for the
coordinate transformation when the poisson factor is included in the prior.
Also add a script which takes a text file with such values and transforms them
into a root TTree which should be loaded into pwaEvidence in the next step.