multinest: Implement coordinate transformation.
Implement the coordinate transformation between the unit hypercube and the
parameter space. In contrast to the previous version of this, the current
implementation uses a prior which follows the poisson factor for the number of
events.

The current version is not finished yet, the transformation according to the
accepted integral matrix is missing as is the modified likelihood
implementation (i.e. without poisson factor). In addition, there is a lot of
debug printout still in which I wanted to save in case it is needed.