multinest: Finish coordinate transformation.
Finish the coordinate transformation by adding the Cholesky decomposition and
using it to bring back interference effects. This is quite messy at the moment
and is probably very incorrect, as the correct mapping between the parameters
as used by DoEval() and the complex normalization matrix with acceptance is not
easy to find out.