amplitude: Change layout in the lorentz factor files.
Change the structure of the files caching the lorentz factors: all the lorentz
factor functions which are coming out of one call to CalcAmpl are now stored
together into one file, drastically reducing the number of files. This also
makes it possible to decide if a lorentz factor function is zero: if the file
is there but the function is not present in the file, the lorentz factor is
zero.