pyInterface: Start fixing scripts and bugs.
Start fixing the user scripts, in this iteration the amplitude calculation and
genPseudoData. Also fix several bugs in the file manager and revise the
interface somewhat. At this state, most scripts are still not operational and
there is quite some commented code in the file manager which needs to be
cleaned up at some point.