pyInterface: Make pwaFit work again.
Make the pwaFit.py script work with the new file manager. This is not done in a
particularly elegant way at the moment, as one has to specify the integral bin
by number instead of actually giving a bin and then have the script magic merge
together the corresponding integral file as needed.