pyInterface: Clean up.
Clean up a lot of code touched in the last few commits. Most of the changes are
of cosmetic nature, an exception is the deWeight.py script, which will no
longer save the weight in the output file. Also fix a few bugs in branches of
the code not explored by the test cases. Note that the pyLint test still fails
for several scripts and the module itself. The latter will hopefully be fixed
soon.