Propagate unit change (MeV -> GeV).
Propagate the change of units from MeV to GeV to several more places, including
many wavelist files, some more scripts and printouts and to the mass dependent
fit. A notable exception are the generator python scripts, because they need
the mass on the command line which gets cumbersome if done with decimal points.