pyInterface: Fix a bug in the file binning logic.
Fix a bug where the in-file binning maps for all the event file types were
added to the binning list as separate bins, leading to duplicate entries in the
bin list. Now, having bins with overlaps triggers a warning and a bin which is
already in the bin list is not added at all.